stock

+++ database/creationTables.sql
@@ -355,7 +355,7 @@ CREATE TABLE `quantitecommande` (
   `idCommande` int NOT NULL,
   `nom` char(100) NOT NULL,
   `emailClient` varchar(100) NOT NULL,
-  `quantite` decimal(2,0) NOT NULL,
+  `quantite` int NOT NULL,
   `prix` decimal(10,2) NOT NULL
 ) ENGINE=InnoDB DEFAULT CHARSET=utf8mb4 COLLATE=utf8mb4_0900_ai_ci;
